requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
promise==2.3
psycopg2==2.8.5
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
